Setup isFavorites

+++ core/src/main/java/io/astronout/core/data/source/local/dao/RemoteKeysDao.kt
@@ -13,7 +13,7 @@ interface RemoteKeysDao {
     suspend fun insertAll(remoteKey: List<RemoteKeys>)
 
     @Query("SELECT * FROM remote_keys WHERE id = :id")
-    suspend fun getRemoteKeysId(id: Int): RemoteKeys?
+    suspend fun getRemoteKeysId(id: Long): RemoteKeys?
 
     @Query("DELETE FROM remote_keys")
     suspend fun clearRemoteKeys()
